backend/requirements.txt
asgiref==3.10.0
Django==5.2.8
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.6.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dotenv==1.2.1
sqlparse==0.5.3
Pillow==10.4.0
